Convert JavaScriptContext trait to class
