+++ src/Medology/Behat/Mink/JavaScriptContext.php
@@ -1,0 +1,155 @@
+<?php namespace Medology\Behat\Mink;
+
+use Behat\Behat\Context\Context;
+use Behat\Behat\Context\Environment\InitializedContextEnvironment;
+use Behat\Behat\Hook\Scope\BeforeScenarioScope;
+use Behat\FlexibleMink\Context\FlexibleContext;
+use Behat\Gherkin\Node\TableNode;
+use Behat\Mink\Exception\ExpectationException;
+use Medology\Behat\GathersContexts;
+use RuntimeException;
+
+/**
+ * {@inheritdoc}
+ */
+class JavaScriptContext implements Context, GathersContexts
+{
+    /** @var FlexibleContext */
+    protected $flexibleContext;
+
+    /**
+     * {@inheritdoc}
+     */
+    public function gatherContexts(BeforeScenarioScope $scope)
+    {
+        $environment = $scope->getEnvironment();
+
+        if (!($environment instanceof InitializedContextEnvironment)) {
+            throw new RuntimeException(
+                'Expected Environment to be ' . InitializedContextEnvironment::class .
+                    ', but got ' . get_class($environment)
+          );
+        }
+
+        if (!$this->flexibleContext = $environment->getContext(FlexibleContext::class)) {
+            throw new RuntimeException('Failed to gather FlexibleContext');
+        }
+    }
+
+    /**
+     * Determines if a javascript variable is set and has a value.
+     *
+     * @Then   the javascript variable :variable should not be null
+     * @param  string               $variable The variable to check.
+     * @throws ExpectationException if the given variable is undefined.
+     */
+    public function assertJavascriptVariableHasAValue($variable)
+    {
+        // Get the value of our variable from javascript
+        $result = $this->flexibleContext->getSession()->evaluateScript('return ' . $variable . ';');
+
+        // If it's null - we failed
+        if ($result === null) {
+            throw new ExpectationException(
+                'The custom variable "' . $variable . '" is null or does not exist.',
+                $this->flexibleContext->getSession()
+            );
+        }
+    }
+
+    /**
+     * Determines if the type of a javascript variable matches a specific type.
+     *
+     * @Then   /^the javascript variable "(?P<variable>.+)" should(?P<not>| not) be type "(?P<type>[\w]+)"$/
+     * @param  string               $variable The variable to evaluate type for.
+     * @param  string|bool          $not      Invert the check? Cast as a boolean based on PHP typecasting.
+     * @param  string               $type     The type to match against.
+     * @throws ExpectationException if the type of the given variable does not match what's expected.
+     */
+    public function assertJavascriptVariableType($variable, $not, $type)
+    {
+        // Get the type of our variable from javascript.
+        $result = $this->flexibleContext->getSession()->evaluateScript('return typeof(' . $variable . ');');
+
+        // If it doesn't match - we failed.
+        if ($result != $type xor $not) {
+            throw new ExpectationException(
+                "The variable \"$variable\" should$not be type $type, but is $result",
+                $this->flexibleContext->getSession()
+            );
+        }
+    }
+
+    /**
+     * Selectively compares two JSON objects.
+     *
+     * @Then   the javascript variable :variableName should have the following contents:
+     * @param  string               $variableName The name of the JS variable to look for.
+     * @param  TableNode            $values       JavaScript variable key-value pair.
+     * @throws ExpectationException if the Javascript variable isn't a match
+     */
+    public function assertJsonContentsOneByOne($variableName, TableNode $values)
+    {
+        $returnedJsonData = $this->flexibleContext->getSession()->evaluateScript(
+            'return JSON.stringify(' . $variableName . ');'
+        );
+        $response = json_decode($returnedJsonData, true);
+
+        foreach ($values->getHash() as $row) {
+            if (!isset($response[$row['key']])) {
+                throw new ExpectationException(
+                    "Expected key \"{$row['key']}\" was not in the JS variable \"{$variableName}\"\n" .
+                        "Actual: $returnedJsonData",
+                    $this->flexibleContext->getSession()
+                );
+            }
+            $expected = $this->getRawOrJson($row['value']);
+            $actual = $this->getRawOrJson($response[$row['key']]);
+
+            if ($actual != $expected) {
+                throw new ExpectationException(
+                    "Expected \"$expected\" in {$row['key']} position but got \"$actual\"",
+                    $this->flexibleContext->getSession()
+                );
+            }
+        }
+    }
+
+    /**
+     * Returns as-is literal inputs (string, int, float), otherwise
+     * returns the JSON encoded output.
+     *
+     * @param  mixed  $value
+     * @return string JSON encoded string
+     */
+    protected function getRawOrJson($value)
+    {
+        if (is_array($value) || is_object($value)) {
+            return json_encode($value);
+        }
+
+        return $value;
+    }
+
+    /**
+     * Asserts that a javascript variable has a specified value.
+     *
+     * @Then   the javascript variable :variableName should have the value of :expectedValue
+     * @param  string               $variableName  This is the name of the variable to be checked.
+     * @param  string               $expectedValue This is the expected value.
+     * @throws ExpectationException If variable value does not match expected value.
+     */
+    public function assertJavascriptVariable($variableName, $expectedValue)
+    {
+        $returnedValue = $this->flexibleContext->getSession()->evaluateScript(
+            'return ' . $variableName . ';'
+        );
+
+        if ($returnedValue != $expectedValue) {
+            throw new ExpectationException(
+                "Expected \"$expectedValue\" but got \"$returnedValue\"",
+                $this->flexibleContext->getSession()
+            );
+        }
+    }
+}
